requirements.txt
GitPython==3.1.47
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMuPDF==1.26.5
PyYAML==6.0.3
Pygments==2.20.0
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
distro==1.9.0
docstring_parser==0.18.0
evaluate==0.4.6
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.13
importlib_resources==6.5.2
jiter==0.14.0
joblib==1.5.3
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.20.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==21.0.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==15.0.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.1
